Report number of variables in scope
It is /very important/ to report a number of variables (numVars) for
larger scopes. If we just say "Nothing", then all variables of all
scopes will be fetched at every stopped event.

Fetching all module level variables and all in the interactive context
is very expensive, so it is great that by reporting the number for
pagination then it doesn't happen at every single break.