Use depth=1 for Module and Global vars
This is important because for very large scopes, it's way too slow to
lookupType every type in the module/global scope with a larger depth (it
even seems to hang, because it is so slow).

TODO: When a variable is forced, it requires all variables to be
re-fetched from scratch. Global variables could indeed change because
some other is forced. However, this is still slow in large programs.

We should consider if it is really all that useful to also display the
structure of global variables. Perhaps just the name (and type) is more
than good enough for global vars. (We can keep the module ones).