Always add entrypoint file to the list of targets
Some cradles don't list all cradles accurately. For example, the
'Default' cradle doesn't list any modules.
We want to support such cradles as well, as it is usually the first
point of contact of beginners with the debugger.
To support this, we add the entrypoint file to the list of targets if
and only if it isn't already a 'Target'.
Moreover, we only do this, if the cradle isn't producing multiple home
unit arguments. Otherwise, we couldn't guess the home unit of the
entryoint file.