Run hdb proxy with runInTerminal
This commit makes the hdb server send a reverse-request to the client
and ask it to run @hdb proxy --port ...@ in a Terminal connected to the
client. From this Terminal, the user can send input to the debuggee and
read its output without additional debugger messages. It is just as if
the program was running in that terminal, but it is just a shim program
which forwards input/output to the debugger that is running the debuggee

Fixes #44