Expand Terms as they the user expands them
The defaultDepth to which we expand Ids is low because otherwise the
performance cost is too great and perceivably slow.

However, the user has to iteractively expand the branches of the
representation tree of variables as they want to view them.

We can leverage this interactivity to overcome the limited depth to
which we expand. Simply: when the user expands a variable, always expand
its corresponding Term a little bit more.

Fixes #97