A fully evaluated may still have children nodes
We were historically not displaying children nodes for fully evaluated
types, but arguably they used to be displayed whole inline, rather than
just the term head.

In any case, the right thing to do is to of course display the nodes for
fully evaluated terms, and keep printing the term head for the specific variable
node.

Fixes #110