feat: Return structured variables for evaluate var
Leverages DAP feature of returning structured variable reference rather
than the evaluation string when the evaluate request is for a single
variable.

This is useful in the Watch context and for interactive exploration in
the REPL, besides the variables pane.

`show <var>` or `print <var>` can be used to reproduce the old behavior
of printing the variable value whole.

Fixes #116
Fixes #108