feat: Custom Debug Instances
This commit introduces the haskell-view-debugger package with a class
for defining custom visualisations for the debugger

It refactors and improves the internals of the debugger variable
inspection code to leverage custom instances when they are available.

We do this by compiling on the fly the method applied to the dictionary
for the type of the value we're inspecting, loading it, and applying it
to the value.

Lastly, we have special logic to decode the return types (VarValue and
VarFields) from their heap-representation Terms.

We get rid of the ad-hoc special logic for displaying strings and/or
boring types and for seqing, resulting in a more uniform and extensible
design.

This commit does not include two things that will follow up:

1. Caching DebugViewInstances with a Core.Map.Type trie map
2. Loading the DebugView class module and built-in instances
   ourselves when the user program does not transitively depend on
   them.

Fixes #47

Co-authored-by: Matthew Pickering <matthewtpickering@gmail.com>