Omit in-memory compilation output
We try to load GHC.Debugger.View.Class and all the orphan instances
modules whenever the package is not part of the transitive closure.

While doing this, we print a lot of compilation information which is
irrelevant to the user.

With this commit, we'll hide it by default.
It is still printed out with -v3.