fix: Never strip interface pragmas & unfoldings
Fixes the step-out test.

Why was it broken? Step-out with do-notation relies on optimisations to
inline the definition of >>= (at least for now).

However, we now loading GHC.Debugger.View.Class in addition to the user
program. Even if the user specified -O1 in their extra GHC args, we
would load GHC.Debugger.View.Class without optimisations (since it
didn't need them).

The issue is that the module interface containing >>= would be stripped
of all pragmas and unfoldings because GHC.Debugger.View.Class didn't
want them. In turn, when the user Main module was loaded, >>= would not
have the unfoldings the user desired. Thus, step-out didn't work.

The fix is to include -fno-ignore-interface-pragmas to make sure we
never discard the interface additional information, and also
-fno-unoptimized-core-for-interpreter to guarantee this isn't undone by
the interpreter not allowing optimisations by default.

Co-authored-by: Matthew Pickering <matthewtpickering@gmail.com>