fix: Pass extraGhcArgs to ALL units
Make sure to parse and use the extraGhcArgs as dynflags which are shared
across all units.

Failing to do this caused a bug with step-out where `>>=` from base was
being loaded without interface pragmas because GHC.Debugger.View.Class
was being compiled with -O0 while the user program was meant to be
compiled with -O1 (extraGhcArgs = -O -fno-unoptimized-core-for-interpreter).

We want to use the same optimization level across all units, and, for
correctness, pass the extra args to all units.

See also the two previous commits for an alternative way to fix this
particular bug, but with downsides of their own and less principled.

Co-authored-by: Matthew Pickering <matthewtpickering@gmail.com>