Implement the 'Program' DSL for DebugView instances
The 'Program' DSL allows a user to communicate with the debugger about
how to display their type. The 'Program' type is evaluated lazily by the
debugger, and contains a query which allows the instance writer to
determine facts about the terms they are displaying.

The main motivating example is to display the already forced prefix of a
list but not to eagerly evaluate the tail.

Fixes #134