fix: Don't fail always on eval exception
Refactors the functions from Debugger.Runtime.Eval to return an
`Either BadEvalStatus _` rather than throwing an exception.

This allows us to recover from these errors in the debugger monad (which
does not have `catch`).

Importantly, we need to return [] when getting an IPE stacktrace throws
an exception. We don't want to crash the debugger just because we failed
to do that.

Really, the exception we're seeing is a bug in the decoding logic.
See GHC#2664: ("Invalid large bitmap size 0")