Display a BCO-based callstack per thread.
On interpreter threads, IPE information is not available because the
stack mixes compiled frames with interpreter frames.

For those threads, such as the debuggee main thread, we can leverage the
interpreter return frames (RetBCOs) to construct a callstack. We can do
this by looking in the RetBCO frame for a BRK_FUN instruction (meaning
the interpreter can stop there), and finding the source location
associated with that BRK_FUN instruction.

We report the stack of BCO continuation frames which do have a BRK_FUN
and for which we find the source location.

Additionally, we lookup the enclosing span surrounding the breakpoint to
span the whole function in the callstack.

Currently, we don't yet report the function names.

Progress towards #107