RemoteExpr: A DSL for evaluation on the remote process
At various points in the execution of the debugger, we need to execute
expressions on the (possibly remote) debuggee process. This was
previously achieved through a sparse and ad-hoc usage of
`evalStmt`, `evalApplication` and `compileExprRemote`.

We did a lot of redundant compilation and loading. E.g. for wrapping
expressions in `IO [...]` or just in general the operations (such as
`listThreads`, `cloneThreadStack`, etc...) got repeatedly compiled and
loaded on the remote process.

We would use a string representation of complex chains of function
calls. This would be impractical to cache.

Instead, we've introduced `RemoteExpr` which allows remote expressions
to be defined compositionally. This allows us to separate function
application from function atoms which have to be loaded on the remote
process. This level of granularity is much more amenable to caching,
though we leave that for a future commit.

We define the `RemoteExpr` DSL using a GADT with all the allowed
remote expr constructs. These include fully-qualified external
variables, direct `ForeignHValue` references, function application,
`pure @IO`, and `(>>=) @IO`. We use `QualifiedDo` to allow `do` notation
on the debugger to stand for `do` notation on the remote process.

We kill most things in the `GHC.Debugger.Runtime.Eval` module, since
they're all superseded by the better `RemoteExpr` abstraction.