Cache compilation of all cacheable expressions.
This delivers on the promise that `RemoteExpr` makes it easy to cache
the things we need to compile and load on the remote process.

We introduce a cache which keeps a `ForeignHValue` for every `String`
which is parsed, compiled, and loaded on the remote process. Using the
`raw` primitives to compile strings is a little bit of a footgun,
because it would be easy to compile and cache raw expression strings
which would never be used again (and thus never freed!?).

One should *always prefer* to use `RemoteExpr` when possible.
Don't use `compileRaw`, nor `compileExprRemote`!

We also introduce `evalExpr` to abstract and log evaluation rather than
using `evalStmt` directly.

Fixes #143